refactor: address e-vite editor review nits
- useInviteOptions: add explicit return type (UseInviteOptions)
- EventInviteManager: aria-pressed on the theme buttons (match the accent
  swatches) so selection isn't conveyed by color alone
- email: include the description in the plain-text alternative, not just HTML
- rename shared/utils/inviteOptions.ts -> invite-options.ts to match the
  kebab-case type sibling (shared/types/invite-options.ts)
- debounce the preview srcdoc (250ms) so typing the message doesn't refetch
  the iframe's font <link> on every keystroke
